+++ packages/language-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vue-vine/language-service
 
+## 1.7.20
+
+### Patch Changes
+
+- Fix compiler for correctly handling kebab-case props in transform phase.
+- Updated dependencies
+  - @vue-vine/compiler@1.7.20
+
 ## 1.7.19
 
 ### Patch Changes