+++ packages/vue-vine/CHANGELOG.md
@@ -1,5 +1,15 @@
 # vue-vine
 
+## 1.7.20
+
+### Patch Changes
+
+- Fix compiler for correctly handling kebab-case props in transform phase.
+- Updated dependencies
+  - @vue-vine/rsbuild-plugin@1.7.19
+  - @vue-vine/rspack-loader@1.7.19
+  - @vue-vine/vite-plugin@1.7.20
+
 ## 1.7.19
 
 ### Patch Changes